Made dashboard page responsive
