+++ client/src/components/Header.js
@@ -35,6 +35,7 @@ function getNavMenu() {
       ))}
       {Auth.loggedIn() && (
         <a
+          href="/"
           className="text-gray-300 hover:bg-gray-700 hover:text-white block px-3 py-2 rounded-md text-base font-medium"
           onClick={() => Auth.logout()}
         >
